refactor(design-system): split Dialog.tsx into focused modules
Dialog.tsx was 1001 lines mixing four separable concerns that its own
section comments already delineated. Now one module each:

- dialogStack.ts — module-global stack registry + useDialogStack
- useFocusTrap.ts — focus trap + DialogInitialFocus
- useBodyScrollLock.ts — reference-counted body scroll lock
- DialogRoot.tsx — context, overlay/content variants, portal root
- DialogParts.tsx — Header/SubHeader/Body/Split/Sidebar/Pane/Footer + the
  Dialog compound export
- ConfirmDialog.tsx — the convenience wrapper

Dialog.tsx remains as a thin re-export assembly, so the design-system
barrel, Dialog.test.tsx, and every consumer import path are unchanged.
Code moved verbatim; max-lines warning cleared.
